+++ problem8/main.cpp
@@ -32,9 +32,9 @@ int main() {
 71636269561882670428252483600823257530420752963450";
     std::vector<int> numbers(strlen(s));
     std::transform(s, s+strlen(s), numbers.begin(), [](char c) { return c - '0'; });
-    long maxProduct = 0;
+    long long maxProduct = 0;
     for(auto it = numbers.begin(); it < numbers.end() - productLength; it++) {
-        long result = 1;
+        long long result = 1;
         for(int i = 0; i < productLength; i++) {
             result *= *(it+i);
         }
